changed cmd.split to schlex.split + made pre-commit changes + created conftest.py to store fixtures
